+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Now alexa spec is splitted into smaller units testing specific functionality
 - Fixed @sys. slots on dialogflow were converted into a different type
+- Fixed issue on path option on the interaction command
+- Fixed issue with platforms undefined
 
 ## [2.1.2] - 2019-05-08